feat(main): add distaceTo C reference function and implement method in glosavector3

+++ Glosa.Net/Core/Interfaces/IVector.cs
@@ -30,7 +30,7 @@ namespace Glosa.Net.Core.Interfaces
         double Magnitude();
         double AngleBetween(T vector);
         void Set(double n);
-        float DistanceTo(T vector);
+        double DistanceTo(T vector);
         T Interpolate(T vector);
         T Min(T[] vectors);
         T Max(T[] vectors);
